remove main and scratch
